requirements.txt
torch>=2.1.0
gymnasium==0.29.1
pettingzoo==1.24.1
stable-baselines3==2.3.0
hydra-core>=1.3.0
omegaconf>=2.3.0
wandb>=0.15.0
umap-learn>=0.5.4
scikit-learn>=1.3.0
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.17.0
numpy>=1.24.0
pytest>=7.4.0
